Restructure project with organized file layout and improved CI/CD
- Organize source code into src/ directory
- Move tests to tests/ directory
- Separate documentation into docs/ and test_reports/
- Update all import statements for new structure
- Enhance GitHub Actions with manual release approval workflow
- Update helper scripts and documentation for new layout
