+++ src/node_editor_view.py
