+++ tests/test_gui_loading_bugs.py
@@ -13,14 +13,14 @@ import unittest
 import json
 
 # Add project root to path
-sys.path.append(os.path.dirname(os.path.abspath(__file__)))
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from PySide6.QtWidgets import QApplication, QGraphicsView, QTextEdit
 from PySide6.QtCore import QTimer, Qt
 
-from node import Node
-from node_graph import NodeGraph
-from flow_format import FlowFormatHandler, load_flow_file
+from src.node import Node
+from src.node_graph import NodeGraph
+from src.flow_format import FlowFormatHandler, load_flow_file
 
 
 class TestGUILoadingBugs(unittest.TestCase):
